+++ notifications-kibana-plugin/public/pages/Emails/components/tables/RecipientGroupsTable.tsx
@@ -16,9 +16,9 @@
 import {
   RecipientGroupItemType,
   TableState,
-} from '../../../../models/interfaces';
+} from '../../../../../models/interfaces';
 import { Component } from 'react';
-import { CoreServicesContext } from '../../../components/coreServices';
+import { CoreServicesContext } from '../../../../components/coreServices';
 import {
   EuiBasicTable,
   EuiButton,
@@ -28,17 +28,17 @@ import {
   EuiTableFieldDataColumnType,
   EuiTableSortingType,
 } from '@elastic/eui';
-import { SORT_DIRECTION } from '../../../../common';
-import { ROUTES } from '../../../utils/constants';
+import { SORT_DIRECTION } from '../../../../../common';
+import { ROUTES } from '../../../../utils/constants';
 import React from 'react';
 import { Criteria } from '@elastic/eui/src/components/basic_table/basic_table';
 import { Pagination } from '@elastic/eui/src/components/basic_table/pagination_bar';
-import { DEFAULT_PAGE_SIZE_OPTIONS } from '../../Notifications/utils/constants';
+import { DEFAULT_PAGE_SIZE_OPTIONS } from '../../../Notifications/utils/constants';
 import {
   ContentPanel,
   ContentPanelActions,
-} from '../../../components/ContentPanel';
-import { ModalConsumer } from '../../../../public/components/Modal';
+} from '../../../../components/ContentPanel';
+import { ModalConsumer } from '../../../../components/Modal';
 import { DeleteSenderModal } from '../modals/DeleteSenderModal';
 
 interface RecipientGroupsTableProps {}
