+++ notifications-kibana-plugin/public/pages/Emails/components/tables/SendersTable.tsx
@@ -13,9 +13,9 @@
  * permissions and limitations under the License.
  */
 
-import { SenderItemType, TableState } from '../../../../models/interfaces';
+import { SenderItemType, TableState } from '../../../../../models/interfaces';
 import { Component } from 'react';
-import { CoreServicesContext } from '../../../components/coreServices';
+import { CoreServicesContext } from '../../../../components/coreServices';
 import {
   EuiBasicTable,
   EuiButton,
@@ -25,17 +25,17 @@ import {
   EuiTableFieldDataColumnType,
   EuiTableSortingType,
 } from '@elastic/eui';
-import { SORT_DIRECTION } from '../../../../common';
-import { ROUTES } from '../../../utils/constants';
+import { SORT_DIRECTION } from '../../../../../common';
+import { ROUTES } from '../../../../utils/constants';
 import React from 'react';
 import { Criteria } from '@elastic/eui/src/components/basic_table/basic_table';
 import { Pagination } from '@elastic/eui/src/components/basic_table/pagination_bar';
-import { DEFAULT_PAGE_SIZE_OPTIONS } from '../../Notifications/utils/constants';
+import { DEFAULT_PAGE_SIZE_OPTIONS } from '../../../Notifications/utils/constants';
 import {
   ContentPanel,
   ContentPanelActions,
-} from '../../../components/ContentPanel';
-import { ModalConsumer } from '../../../../public/components/Modal';
+} from '../../../../components/ContentPanel';
+import { ModalConsumer } from '../../../../components/Modal';
 import { DeleteSenderModal } from '../modals/DeleteSenderModal';
 
 interface SendersTableProps {}
